fix icon badge
