+++ libs/remix-ui/vertical-icons-panel/src/lib/components/Icon.tsx
@@ -3,9 +3,10 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import VerticalIconsContextMenu from '../vertical-icons-context-menu'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import React, { Fragment, SyntheticEvent, useEffect, useReducer, useRef, useState } from 'react'
+import React, { Dispatch, Fragment, SyntheticEvent, useEffect, useReducer, useRef, useState } from 'react'
 import { VerticalIcons } from 'libs/remix-ui/vertical-icons-panel/types/vertical-icons-panel'
 import Badge from './Badge'
+import { iconBadgeReducer, IconBadgeReducerAction } from '../reducers/iconBadgeReducer'
 
   interface IconProps {
     verticalIconPlugin: VerticalIcons
@@ -13,7 +14,6 @@ import Badge from './Badge'
     contextMenuAction: (evt: any, profileName: string, documentation: string) => void
     addActive: (profileName: string) => void
     removeActive: () => void
-    badgeStatus?: BadgeStatus
   }
 
 export interface IconStatus {
@@ -41,14 +41,21 @@ export interface IconProfile {
     tooltip?: string
   }
 
+const initialState = {
+  text: '',
+  key: '',
+  title: '',
+  type: '',
+  pluginName: ''
+}
+
 function Icon ({
   profile,
   verticalIconPlugin,
   contextMenuAction,
   addActive,
-  removeActive,
+  removeActive
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  badgeStatus
 }: IconProps) {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { tooltip, displayName, name, kind, icon, documentation } = profile
@@ -59,6 +66,7 @@ function Icon ({
   const [links, setLinks] = useState<{ Documentation: string, CanDeactivate: boolean }>(
       {} as { Documentation: string, CanDeactivate: boolean }
   )
+  const [badgeStatus, dispatchStatusUpdate] = useReducer(iconBadgeReducer, initialState)
   // @ts-ignore
   const [pageX, setPageX] = useState<number>(null)
   // @ts-ignore
@@ -84,6 +92,14 @@ function Icon ({
   function closeContextMenu () {
     setShowContext(false)
   }
+
+  useEffect(() => {
+    verticalIconPlugin.on(name, 'statusChanged', (iconStatus: IconStatus) => {
+      iconStatus.pluginName = name
+      const action: IconBadgeReducerAction = { type: name, payload: { status: iconStatus, verticalIconPlugin: verticalIconPlugin } }
+      dispatchStatusUpdate(action)
+    })
+  }, [])
 
   return (
     <Fragment>
@@ -114,7 +130,7 @@ function Icon ({
         <img className="remixui_image" src={icon} alt={name} />
         { badgeStatus && badgeStatus.pluginName === name ? (
           <Badge
-            badgeStatus={badgeStatus!}
+            badgeStatus={badgeStatus}
           />) : null }
       </div>
       {showContext ? (
